Move Firebase commands to separate file, add support for querying single problem
